docker_compose_hw/hw/dj_crud_stocks/requirements.txt
django~=4.0.4
djangorestframework~=3.13.1
psycopg2-binary
gunicorn
django-filter~=21.1